refactor json path argument helper
